Projets_Data_Engineering/API_OpenFoodFact/requirements.txt
aniso8601==9.0.1
autopep8==1.5.6
bcrypt==3.2.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.1.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.23.0
pip==20.2.4
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.20
PyJWT==2.0.1
pymongo==3.5.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.24.0
ruamel-yaml==0.15.87
six==1.15.0
toml==0.10.2
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
wheel==0.35.1
WTForms==2.3.3
